+++ python/GafferSceneUI/HierarchyView.py
@@ -238,7 +238,7 @@ class HierarchyView( GafferUI.NodeSetEditor ) :
 
 		selection = ContextAlgo.getSelectedPaths( self.getContext() )
 		with Gaffer.Signals.BlockedConnection( self.__selectionChangedConnection ) :
-			self.__pathListing.setSelection( selection, scrollToFirst=True, expandNonLeaf=False )
+			self.__pathListing.setSelection( selection, scrollToFirst=True )
 
 GafferUI.Editor.registerType( "HierarchyView", HierarchyView )