+++ Changes.md
@@ -100,6 +100,7 @@ Breaking Changes
   - Removed `availableColorSpaces()` and `availableRoles()` methods.
   - Deprecated `context` plug.
 - OpenColorIO : Changed default config.
+- PathListingWidget : Removed `expandNonLeaf` argument from `setSelection()` and `setSelectedPaths()` methods.
 - Subprocess32 : Removed Python module.
 - Six : Removed Python module.
 - Gaffer : Class constructors are now declared explicit disabling implicit conversions.